+++ crates/gen/src/lib.rs
@@ -9,6 +9,7 @@ mod delegate;
 mod r#enum;
 mod function;
 mod gen;
+mod guid;
 mod handle;
 mod helpers;
 mod interface;
@@ -16,6 +17,7 @@ mod interface_info;
 mod iterator;
 mod matrix3x2;
 mod matrix4x4;
+mod name;
 mod ntstatus;
 mod object;
 mod propertykey;
@@ -24,17 +26,21 @@ mod pwstr;
 mod r#struct;
 mod timespan;
 mod to_ident;
+mod tree;
 mod vector2;
 mod vector3;
 mod vector4;
-mod guid;
-mod name;
-mod tree;
+mod win32;
+mod winrt;
 
+// TODO: These public things are mostly/all used by the implement macro
+// move that logic into here so the macro crate is just forwarding to the gen crate
 pub use gen::*;
 pub use helpers::*;
 pub use interface_info::*;
+pub use name::*;
 pub use to_ident::*;
+pub use tree::*;
 
 use bool32::*;
 use bstr::*;
@@ -44,6 +50,7 @@ use com_interface::*;
 use constant::*;
 use delegate::*;
 use function::*;
+use guid::*;
 use handle::*;
 use interface::*;
 use iterator::*;
@@ -63,6 +70,5 @@ use timespan::*;
 use vector2::*;
 use vector3::*;
 use vector4::*;
-use guid::*;
-pub use name::*;
-pub use tree::*;
+pub use win32::*;
+pub use winrt::*;
